update calibrate, move name manager and symbol helper from tfm_base into tfm_pass
