+++ cvm/quantization/tfm_base.py
@@ -1,4 +1,5 @@
 from sym_utils import *
+from tfm_pass import * 
 
 import numpy as np
 
@@ -141,134 +142,6 @@ def apply_pass(pass_t, **updates):
             kwargs[n][ret.attr('name')] = kwargs[n][op.attr('name')]
         return ret
     return wrapper
-
-_NoneName = object()
-class N(object):
-    _global_name = _NoneName
-    _count = {}
-    _name_manager = {}
-
-    @staticmethod
-    def n(name=""):
-        assert N._global_name != _NoneName, \
-                "register name manager first please"
-        if name not in N._name_manager:
-            N._name_manager[name] = 0
-        _n = "mrt"
-        if N._global_name:
-            _n += "_" + N._global_name
-        if name != "":
-            _n += "_" + name
-        _n += "_%d" % N._name_manager[name]
-        N._name_manager[name] += 1
-        return _n
-
-    @staticmethod
-    def _set_global(name):
-        if name not in N._count:
-            N._count[name] = 0
-        else:
-            name += str(N._count[name])
-        N._global_name = name
-
-    @staticmethod
-    def register_nm(name):
-        def wrapper(pass_f):
-            def run(symbol, params, *args, **kwargs):
-                old_name = N._global_name
-                N._set_global(name)
-                N._count[name] += 1
-                ret = pass_f(symbol, params, *args, **kwargs)
-                N._global_name = old_name
-                return ret
-            return run
-        return wrapper
-
-# === symbol helper ===
-
-@N.register_nm("gv")
-def graph_validate(symbol, params):
-    def _name_replace(op, params, graph):
-        name, op_name = op.attr('name'), op.attr('op_name')
-        childs, attr = sym_iter(op.get_children()), op.list_attr()
-        if is_inputs(op, params):
-            assert "data" not in graph, "multiple inputs"
-            op = mx.sym.var("data", attr=attr)
-        elif is_params(op, params):
-            op = mx.sym.var(N.n("params"), attr=attr)
-            params[op.attr('name')] = params[name]
-        elif childs is None:
-            op = get_mxnet_op(op_name)(name=N.n(op_name), **attr)
-        else:
-            op = get_mxnet_op(op_name)(*childs, name=N.n(op_name), **attr)
-        return op
-    sym, params = topo_visit_transformer(symbol, params, _name_replace)
-
-    new_params = {s.attr('name'):params[s.attr('name')] \
-            for s in topo_sort(sym) if is_params(s, params)}
-
-    return sym, new_params
-
-@N.register_nm("fc")
-def fuse_constant(symbol, params):
-    def _impl(op, params, graph):
-        name, op_name = op.attr('name'), op.attr('op_name')
-        childs, attr = sym_iter(op.get_children()), op.list_attr()
-        if is_var(op, params):
-            pass
-        elif childs is None:
-            params[name] = get_nd_op(op_name)(**attr)
-            op = mx.sym.var(name, shape=params[name].shape)
-        elif all([is_params(c, params) for c in childs]):
-            in_params = [params[c.attr('name')] for c in childs]
-            params[name] = get_nd_op(op_name)(*in_params, **attr)
-            op = mx.sym.var(name, shape=params[name].shape)
-        return op
-    return topo_visit_transformer(symbol, params, _impl)
-
-@N.register_nm("ais")
-def attach_input_shape(symbol, params, input_shape):
-    def _impl(op, params, graph):
-        if is_inputs(op, params):
-            op = mx.sym.var(op.attr('name'), shape=input_shape)
-        return op
-    return topo_visit_transformer(symbol, params, _impl)
-
-def infer_shape(symbol, params, input_shape=None):
-    infer_shapes = {}
-    def _impl(op, params, graph):
-        name, op_name = op.attr('name'), op.attr('op_name')
-        _, oshp, _ = op.infer_shape()
-
-        if is_params(op, params):
-            if oshp is None:
-                oshp = [params[name].shape]
-                op = mx.sym.var(name, shape=oshp[0])
-            assert params[name].shape == oshp[0], \
-                    "Parameter %s's shape %s is inconsistent with \
-                    params dict %s" % (name, oshp[0], params[name].shape)
-        elif is_inputs(op, params):
-            if input_shape is None:
-                assert oshp is not None, "It seems that graph doesn't set \
-                        input_shape, please invoke attach_input_shape first."
-            else:
-                oshp = [input_shape]
-                op = mx.sym.var(name, shape=oshp[0])
-        infer_shapes[name] = oshp
-        return op
-    topo_visit_transformer(symbol, params, _impl)
-    return infer_shapes
-
-def _collect_attribute(op, **kwargs):
-    attr_name, func = kwargs['attr_name'], kwargs['func']
-    func(op.attr(attr_name))
-    return op
-
-def collect_op_names(symbol, params):
-    op_names = set()
-    _ = topo_visit_transformer(symbol, params, _collect_attribute,
-            attr_name='op_name', func=op_names.add)
-    return op_names
 
 # === symbol pass ==
