+++ cvm/quantization/transformer.py
@@ -2,6 +2,7 @@ from tfm_base import *
 import tfm_ops
 import cvm_op
 from sym_utils import *
+from tfm_pass import *
 
 def init(symbol, params, input_shape=None):
     sym, params = graph_validate(symbol, params)
@@ -36,3 +37,7 @@ if __name__ == "__main__":
     print (calculate_ops(sym, params))
     with open(os.path.expanduser("~/tvm-cvm/data/tmp_v2.json"), "w") as fout:
         fout.write(sym.tojson())
+    data_iter_func = ds.data_iter('imagenet', 1, input_size=224)
+    data, _ = data_iter_func()
+    th_dict = sym_calibrate(sym, params, data, old_ths=None, lambd=None)
+
